Initial move to GResource
The main reason for this is for the built in gtk magic that allows
usage of multiple menu files depending on the platform.

This work is not complete but it can be expanded to use gear-menus on
some Linux DEs and already uses a proper menu on OSX and Windows.